split default templates in standalone files
