+++ lib/templates/bootstrap/index.js
@@ -1,0 +1,7 @@
+module.exports = {
+  textbox: require('./textbox'),
+  checkbox: require('./checkbox'),
+  select: require('./select'),
+  datepicker: require('./datepicker'),
+  struct: require('./struct')
+};
